pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kestrel-evals"
version = "0.1.0"
description = "Minimal LLM evaluation harness (Phase 1: OpenAI)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
openai = ["openai>=1.40.0"]

[project.scripts]
kestrel-evals = "kestrel_evals.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
